src/ms_potts/requirements.txt
fastapi
gradio
requests
sentence-transformers
google-generativeai
uvicorn
pydantic
python-dotenv
